% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.densemlp_fit}
\alias{autoplot.densemlp_fit}
\title{Plot training history}
\usage{
\method{autoplot}{densemlp_fit}(object, ...)
}
\arguments{
\item{object}{A fitted \code{densemlp_fit} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot training history
}
