% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.mlp_fit}
\alias{autoplot.mlp_fit}
\title{Plot training history}
\usage{
\method{autoplot}{mlp_fit}(object, ...)
}
\arguments{
\item{object}{A fitted \code{mlp_fit} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot training history
}
