% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mlp_metrics}
\alias{mlp_metrics}
\title{Compute task-aware metrics}
\usage{
mlp_metrics(truth, estimate, task = NULL, prob = NULL)
}
\arguments{
\item{truth}{Ground truth values.}

\item{estimate}{Predicted classes for classification or numeric predictions
for regression.}

\item{task}{Optional task override.}

\item{prob}{Optional class probabilities.}
}
\value{
A named list of metrics.
}
\description{
Compute task-aware metrics
}
