% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.densemlp}
\alias{plot.densemlp}
\title{Plot training history}
\usage{
\method{plot}{densemlp}(x, ...)
}
\arguments{
\item{x}{A fitted \code{densemlp} object (single model; ensembles are not
supported since members don't share an epoch axis).}

\item{...}{Additional arguments passed to \code{\link[graphics:matplot]{graphics::matplot()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot training history
}
