% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot.densemlp_importance}
\alias{plot.densemlp_importance}
\title{Plot permutation importance}
\usage{
\method{plot}{densemlp_importance}(x, top = 20L, ...)
}
\arguments{
\item{x}{A \code{densemlp_importance} object.}

\item{top}{Number of top features to show.}

\item{...}{Passed to \code{\link[graphics:barplot]{graphics::barplot()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot permutation importance
}
