% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.densemlp_fit}
\alias{print.densemlp_fit}
\title{Print a fitted dense multilayer perceptron}
\usage{
\method{print}{densemlp_fit}(x, ...)
}
\arguments{
\item{x}{A fitted \code{densemlp_fit} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fitted dense multilayer perceptron
}
