% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mlp_fit}
\alias{print.mlp_fit}
\title{Print a fitted MLP}
\usage{
\method{print}{mlp_fit}(x, ...)
}
\arguments{
\item{x}{A fitted \code{mlp_fit} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fitted MLP
}
