% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{densemlp_metrics}
\alias{densemlp_metrics}
\title{Compute densemlp metrics}
\usage{
densemlp_metrics(truth, estimate, task = NULL, prob = NULL)
}
\arguments{
\item{truth}{Ground truth values.}

\item{estimate}{Predicted classes for classification or numeric predictions
for regression.}

\item{task}{Optional task override. When \code{NULL}, the task is inferred from
\code{truth} and \code{estimate}.}

\item{prob}{Optional class probabilities.}
}
\value{
A named list of metrics.
}
\description{
Compute densemlp metrics
}
