% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{perm_importance}
\alias{perm_importance}
\title{Permutation variable importance}
\usage{
perm_importance(object, new_data, truth, metric = NULL, seed = object$seed)
}
\arguments{
\item{object}{A fitted \code{mlp_fit} object.}

\item{new_data}{Evaluation data frame.}

\item{truth}{Ground truth outcome values.}

\item{metric}{Metric name. Defaults to accuracy for classification and RMSE
for regression.}

\item{seed}{Random seed used for shuffling.}
}
\value{
An \code{mlp_importance} object.
}
\description{
Permutation variable importance
}
