% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot.densemlp_importance}
\alias{plot.densemlp_importance}
\title{Plot permutation importance}
\usage{
\method{plot}{densemlp_importance}(x, ...)
}
\arguments{
\item{x}{A \code{densemlp_importance} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot permutation importance
}
