% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_history}
\alias{plot_history}
\title{Plot training history}
\usage{
plot_history(object)
}
\arguments{
\item{object}{A fitted \code{densemlp_fit} object.}
}
\value{
A ggplot object.
}
\description{
Plot training history
}
